improve state-sync tests module
